+++ src/routes/oauth/oauth.js
@@ -35,7 +35,7 @@ router
 
         if (req.query.forceNewLogin) {
             let baseUrl = config.url
-            let backToHereUrl = baseUrl + '/oauth/site/' + req.site.id + '/login?' + (req.query.useOauth ? 'useOauth=' + req.query.useOauth : '') + '&redirectUrl=' + req.query.redirectUrl
+          let backToHereUrl = baseUrl + '/oauth/site/' + req.site.id + '/login?' + (req.query.useOauth ? 'useOauth=' + req.query.useOauth : '') + '&redirectUrl=' + encodeURIComponent(req.query.redirectUrl)
             backToHereUrl = encodeURIComponent(backToHereUrl)
             let url = baseUrl + '/oauth/site/' + req.site.id + '/logout?redirectUrl=' + backToHereUrl;
 
@@ -264,7 +264,6 @@ router
 
         let which = req.query.useOauth || 'default';
         let siteOauthConfig = (req.site && req.site.config && req.site.config.oauth && req.site.config.oauth[which]) || {};
-        ;
 
         let authServerUrl = siteOauthConfig['auth-server-url'] || config.authorization['auth-server-url'];
         let authServerGetUserPath = siteOauthConfig['auth-server-logout-path'] || config.authorization['auth-server-logout-path'];